requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
boxmot==13.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
cython_bbox==0.1.5
deep-sort-realtime==1.3.2
filelock==3.18.0
filterpy==1.4.5
fonttools==4.58.1
fsspec==2025.5.1
ftfy==6.3.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lapx==0.5.11.post1
lightning-utilities==0.14.3
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.3
motmetrics==1.4.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycocotools==2.0.10
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.29.1
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
soupsieve==2.7
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.7.2
torchvision==0.20.1+cu121
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.146
ultralytics-thop==2.0.14
urllib3==2.4.0
wandb==0.20.1
wcwidth==0.2.13
win32_setctime==1.2.0
xmltodict==0.14.2
yacs==0.1.8
